Fix mobile firefox inputs
