+++ src/use-position-fixed.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { isSafari } from './use-prevent-scroll';
+import { isSafari } from './browser';
 
 let previousBodyPosition: Record<string, string> | null = null;
